// apps/marketing/src/content/types.ts
export type FooterContent = {
  description: string;
  links: {
    title: string;
    items: { href: string; label: string }[];
  }[];
  legal: { href: string; label: string }[];
  cta: {
    title: string;
    description: string;
    action: string;
    href: string;
  };
  copyright: string;
};